+++ pmd/src/test/java/net/sourceforge/pmd/lang/java/rule/XPathRuleTest.java
@@ -128,6 +128,49 @@ import org.junit.Test;
          assertEquals(1, nodes.size());
      }
 
+     /**
+      * Following sibling check: See https://sourceforge.net/p/pmd/bugs/1209/
+      * @throws Exception any error
+      */
+     @Test
+     public void testFollowingSibling() throws Exception {
+         String SOURCE = "public class dummy {\n" +
+                 "  public String toString() {\n" +
+                 "    String test = \"bad example\";\n" +
+                 "    test = \"a\";\n" +
+                 "    return test;\n" +
+                 "  }\n" +
+                 "}";
+         LanguageVersion language = Language.JAVA.getDefaultVersion();
+         ParserOptions parserOptions = language.getLanguageVersionHandler().getDefaultParserOptions();
+         Parser parser = language.getLanguageVersionHandler().getParser(parserOptions);
+         ASTCompilationUnit cu = (ASTCompilationUnit)parser.parse("test", new StringReader(SOURCE));
+         RuleContext ruleContext = new RuleContext();
+         ruleContext.setLanguageVersion(language);
+
+         String xpath = "//Block/BlockStatement/following-sibling::BlockStatement";
+
+         // XPATH version 1.0
+         XPathRuleQuery xpathRuleQuery = new JaxenXPathRuleQuery();
+         xpathRuleQuery.setXPath(xpath);
+         xpathRuleQuery.setProperties(new HashMap<PropertyDescriptor<?>, Object>());
+         xpathRuleQuery.setVersion(XPathRuleQuery.XPATH_1_0);
+         List<Node> nodes = xpathRuleQuery.evaluate(cu, ruleContext);
+         assertEquals(2, nodes.size());
+         assertEquals(4, nodes.get(0).getBeginLine());
+         assertEquals(5, nodes.get(1).getBeginLine());
+
+         // XPATH version 2.0
+         xpathRuleQuery = new SaxonXPathRuleQuery();
+         xpathRuleQuery.setXPath(xpath);
+         xpathRuleQuery.setProperties(new HashMap<PropertyDescriptor<?>, Object>());
+         xpathRuleQuery.setVersion(XPathRuleQuery.XPATH_2_0);
+         nodes = xpathRuleQuery.evaluate(cu, ruleContext);
+         assertEquals(2, nodes.size());
+         assertEquals(4, nodes.get(0).getBeginLine());
+         assertEquals(5, nodes.get(1).getBeginLine());
+     }
+
      private static final String TEST1 =
              "public class Foo {" + PMD.EOL +
              " int a;" + PMD.EOL +