+++ pmd/src/site/markdown/changelog.md
@@ -7,6 +7,7 @@
 * Fixed [bug #1181]: unused import false positive if used as parameter in javadoc only.
 * Fixed [bug #1192]: Ecmascript fails to parse this operator " ^= "
 * Fixed [bug #1206]: SummaryHTMLRenderer always shows suppressed warnings/violations
+* Fixed [bug #1209]: XPath 2.0 following-sibling incorrectly includes context node
 * Fixed [bug #1211]: PMD is failing with NPE for rule UseIndexOfChar while analyzing Jdk 8 Lambda expression
 * Fixed [bug #1214]: UseCollectionIsEmpty misses some usage
 * Fixed [bug #1215]: AvoidInstantiatingObjectsInLoops matches the right side of a list iteration loop
@@ -19,6 +20,7 @@
 [bug #1181]: https://sourceforge.net/p/pmd/bugs/1181/
 [bug #1192]: https://sourceforge.net/p/pmd/bugs/1192/
 [bug #1206]: https://sourceforge.net/p/pmd/bugs/1206/
+[bug #1209]: https://sourceforge.net/p/pmd/bugs/1209/
 [bug #1211]: https://sourceforge.net/p/pmd/bugs/1211/
 [bug #1214]: https://sourceforge.net/p/pmd/bugs/1214/
 [bug #1215]: https://sourceforge.net/p/pmd/bugs/1215/